Refactoring and removing unused resources
